Typescriptification
Straight dabbing on the haters this thing is ts now
