+++ package.json
@@ -4,7 +4,13 @@
   "description": "Discord bot for shooting_the_breeze.java",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc --skipLibCheck",
+    "start": "nodemon ./dist/index.js",
+    "prod": "npm run build && npm run start",
+    "format": "prettier-standard '**/*.ts'",
+    "deploy": "npm run build && node ./dist/deploy-commands.js",
+    "clean": "rimraf dist/"
   },
   "repository": {
     "type": "git",
@@ -22,12 +28,35 @@
   },
   "homepage": "https://github.com/jswaneck/BotReginald#readme",
   "dependencies": {
-    "@discordjs/builders": "^0.6.0",
     "@discordjs/rest": "^0.1.0-canary.0",
-    "discord-api-types": "^0.22.0",
-    "discord.js": "^13.1.0"
+    "class-transformer": "0.4.0",
+    "class-validator": "0.13.1",
+    "discord.js": "13.1.0",
+    "discord.js-multilingual-utils": "1.7.2",
+    "discord.js-rate-limiter": "1.2.2",
+    "express": "4.17.1",
+    "express-promise-router": "4.1.0",
+    "filesize": "8.0.0",
+    "node-fetch": "2.6.1",
+    "node-schedule": "2.0.0",
+    "pino": "6.13.1",
+    "pino-pretty": "6.0.0",
+    "pm2": "^5.1.1",
+    "reflect-metadata": "^0.1.13",
+    "remove-markdown": "0.3.0",
+    "rimraf": "^3.0.2",
+    "rm": "^0.1.8"
   },
   "devDependencies": {
-    "eslint": "^7.32.0"
+    "@types/express": "4.17.13",
+    "@types/node": "16.7.2",
+    "@types/node-fetch": "2.5.12",
+    "@types/node-schedule": "1.3.2",
+    "@types/pino": "6.3.11",
+    "@types/remove-markdown": "0.3.1",
+    "prettier": "^2.3.2",
+    "ts-node-dev": "^1.1.8",
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.2"
   }
 }
