+++ tsconfig.json
@@ -1,0 +1,40 @@
+{
+    "compilerOptions": {
+        "target": "es2020",
+        "module": "commonjs",
+        "lib": ["es2021"],
+        "declaration": true,
+        "sourceMap": true,
+        "outDir": "./dist",
+        "rootDir": "./src",
+        "strict": false,
+        "baseUrl": "./dist",
+        "moduleResolution": "node",
+        "esModuleInterop": true,
+        "experimentalDecorators": true,
+        "emitDecoratorMetadata": true,
+        "skipLibCheck": true,
+        "forceConsistentCasingInFileNames": true,
+        "resolveJsonModule": true,
+
+    },
+    "types": [
+      // add node as an option
+      "node",
+      "reflect-metadata"
+    ],
+    "include": [
+        "src/*",
+        "src/commands/*",
+        "config.json"
+    ],
+    "typeRoots": [
+      // add path to @types
+      "./node_modules/@types"
+    ],
+    "exclude": [
+        "dist", 
+        "node_modules", 
+        "tests"
+    ]
+}
